Add initial BottomTabNavigator
